+++ src/app/watch.py
@@ -1,14 +1,13 @@
-import os
+import subprocess
 from pathlib import Path
 from typing import Annotated
 
 import typer
-from livereload import Server
 from loguru import logger
-from weasyprint import html
+from watchfiles import run_process
 
 from app.constants import ColorScheme
-from app.resume import build_page, __file__ as resume_module_file, get_html_path
+from app.build import build, build_page, __file__ as build_module_file, get_html_path
 
 
 cli = typer.Typer()
@@ -18,20 +17,17 @@ cli = typer.Typer()
 def page(
     color: Annotated[ColorScheme, typer.Option(help="Render with color scheme.")] = ColorScheme.light,
 ):
-    def _build():
-       build_page(color)
-
     html_path = get_html_path()
-    _build()
-    server = Server()
-    logger.debug(f"Adding README.md to watchers")
+    static_path = Path("static")
     readme_path = Path("README.md")
-    server.watch(readme_path, func=_build)
-    etc_path = Path("etc/css")
-    for css_path in etc_path.glob("*.css"):
-        logger.debug(f"Adding {css_path} to watchers")
-        server.watch(str(css_path), func=_build)
-    logger.debug(f"Adding {resume_module_file} to watchers")
-    server.watch(resume_module_file, func=_build)
 
-    server.serve(default_filename=str(html_path), open_url_delay=0.25, live_css=False)
+    watch_paths = [build_module_file, html_path, static_path, readme_path]
+
+    logger.debug("Triggering initial build")
+    build_page(color, debug=True)
+
+    logger.debug("Opening browser to page")
+    subprocess.run(["explorer.exe", str(html_path)])
+
+    logger.debug(f"Watching for changes in {watch_paths}")
+    run_process(*watch_paths, target=build_page, args=(color,), kwargs=dict(debug=True))